You can enter number with comma now.

+++ converter/Converter.py
@@ -1,2 +1,2 @@
-f = int(input("Enter the temperature in °F: "))
+f = float(input("Enter the temperature in °F: "))
 print(f, "°F is ", (f - 32) * 5 / 9, "°C.")
